# internal/api/handlers/auth_handler.go
package handlers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/davidreturns08/fatch/internal/database"
	"github.com/davidreturns08/fatch/internal/services"
	"github.com/davidreturns08/fatch/internal/types"
)

var authService services.AuthService

// @Summary Change password for a user
// @Tags auth
// @Accept json
// @Produce json
// @Param request body database.ChangePasswordParams true "Request body for changing user password"
// @Success 200 {object} types.SuccessResponse
// @Failure 400 {object} types.ErrorResponse
// @Failure 412 {object} types.ErrorResponse
// @Failure 500 {object} types.ErrorResponse
// @Router /auth/passwd [post]
func ChangePassword(w http.ResponseWriter, r *http.Request) *types.ErrorDetails {
	ctx := r.Context()
	var params database.ChangePasswordParams

	if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
		return types.ReturnJSON(w, types.BadRequestErrorResponse("Invalid JSON data"))
	}

	// VALIDATE INPUT
	// Missing fields validation
	if params.OldPasswd == "" || params.NewPasswd == "" {
		return types.ReturnJSON(w,
			types.BadRequestErrorResponse("No empty fields allowed"),
		)
	}

	// password length validation
	if len(params.OldPasswd) < 8 || len(params.NewPasswd) < 8 {
		return types.ReturnJSON(w,
			types.PreconditionFailedErrorResponse("Password must be at least 8 characters long"),
		)
	}

	// don't allow same passwords
	if params.OldPasswd == params.NewPasswd {
		return types.ReturnJSON(w, types.PreconditionFailedErrorResponse("Old and new password fields must be different"))
	}

	// call service to change password
	errResponse, err := authService.ChangePassword(ctx, params)
	if err != nil {
		types.ReturnJSON(w, errResponse)
		return &types.ErrorDetails{
			Message: "Failed to change password",
			Trace:   err,
		}
	}

	// Send success response
	return types.ReturnJSON(w, types.OKResponse("Password changed successfully", nil))
}

// @Summary Verify user login attempt
// @Tags auth
// @Accept json
// @Produce json
// @Param request body database.VerifyPasswordParams true "Request body for verifying user login attempt"
// @Success 200 {object} types.SuccessResponse
// @Failure 400 {object} types.ErrorResponse
// @Failure 500 {object} types.ErrorResponse
// @Router /auth/verify [post]
func VerifyPassword(w http.ResponseWriter, r *http.Request) *types.ErrorDetails {
	ctx := r.Context()
	var params database.VerifyPasswordParams

	if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
		return types.ReturnJSON(w, types.BadRequestErrorResponse("Invalid JSON data"))
	}

	// VALIDATE INPUT
	// Missing fields validation: Can sign in with email or username
	if (params.Email == "" && params.Username == "") || params.Passwd == "" {
		return types.ReturnJSON(w,
			types.BadRequestErrorResponse("No empty fields allowed"),
		)
	}

	// call service to create user
	response, errResponse, err := authService.VerifyPassword(ctx, params)
	if err != nil {
		types.ReturnJSON(w, errResponse)
		return &types.ErrorDetails{
			Message: "Failed to verify password",
			Trace:   err,
		}
	}

	// get user info
	user, errResponse, err := userService.GetUserById(ctx, database.GetUserByIdParams{
		UserID: response.UserID,
	})
	if err != nil {
		types.ReturnJSON(w, errResponse)
		return &types.ErrorDetails{
			Message: "Failed to get user info",
			Trace:   err,
		}
	}

	responseMsg := fmt.Sprintf("%s logged in successfully", user.Username)

	// return success response
	return types.ReturnJSON(w, types.OKResponse(responseMsg, response))
}